feat(fe)!: protect all routes from unauthorised users (#25)
adds middleware to protect all routes from unauthorised users.

moves the login screen from being a subscreen of `/` to a seperate route `/login`.
