+++ src/app/_components/screens/login-screen.tsx
@@ -1,12 +1,18 @@
-"use client";
+"use client"
 
-import type React from "react";
-import { useState } from "react";
-import { Heart, LogIn, Loader2 } from "lucide-react";
-import Image from "next/image";
-import Link from "next/link";
+import { useState } from "react"
+import { Heart, Loader2 } from "lucide-react"
+import Image from "next/image"
+import { signIn } from "next-auth/react"
 
 export default function LoginScreen() {
+  const [isLoading, setIsLoading] = useState(false)
+
+  const handleGoogleSignIn = async () => {
+    setIsLoading(true)
+    await signIn("google", { callbackUrl: "/" })
+  }
+
   return (
     <div className="flex h-full flex-col items-center justify-center bg-[#1a1a1a] p-6">
       <div className="mb-6 flex h-16 w-16 items-center justify-center rounded-full bg-[#c1ff72]">
@@ -14,9 +20,7 @@ export default function LoginScreen() {
       </div>
 
       <h1 className="mb-2 text-3xl font-bold text-[#f3f3f3]">BiggMarket</h1>
-      <p className="mb-8 text-center text-sm text-[#a9a9a9]">
-        Swap, don&apos;t shop. Reduce waste, find treasures.
-      </p>
+      <p className="mb-8 text-center text-sm text-[#a9a9a9]">Swap, don&apos;t shop. Reduce waste, find treasures.</p>
 
       <div className="relative mb-6 flex w-full items-center justify-center">
         <div className="absolute inset-0 flex items-center">
@@ -28,25 +32,26 @@ export default function LoginScreen() {
       </div>
 
       {/* Google signin button */}
-      <Link
-        href="/api/auth/signin"
+      <button
+        onClick={handleGoogleSignIn}
+        disabled={isLoading}
         className="mb-4 flex w-full items-center justify-center rounded-lg bg-[#242424] py-3 font-semibold text-[#f3f3f3]"
       >
-        <Image
-          src="https://upload.wikimedia.org/wikipedia/commons/thumb/c/c1/Google_%22G%22_logo.svg/768px-Google_%22G%22_logo.svg.png"
-          alt="Google"
-          width={20}
-          height={20}
-          className="mr-2 h-5 w-5"
-          draggable={false}
-        />
-        Sign in
-      </Link>
-
-      <p className="text-sm text-[#a9a9a9]">
-        Don&apos;t have an account?{" "}
-        <button className="font-semibold text-[#c1ff72]">Sign up</button>
-      </p>
+        {isLoading ? (
+          <Loader2 className="mr-2 h-5 w-5 animate-spin" />
+        ) : (
+          <Image
+            src="https://upload.wikimedia.org/wikipedia/commons/thumb/c/c1/Google_%22G%22_logo.svg/768px-Google_%22G%22_logo.svg.png"
+            alt="Google"
+            width={20}
+            height={20}
+            className="mr-2 h-5 w-5"
+            draggable={false}
+          />
+        )}
+        {isLoading ? "Signing in..." : "Sign in with Google"}
+      </button>
     </div>
-  );
+  )
 }
+
