+++ src/app/profile/[userId]/page.tsx
@@ -12,7 +12,8 @@ import {
 } from "lucide-react";
 import Image from "next/image";
 import AppShell from "@/components/app-shell";
-import { useRouter } from "next/navigation";
+import { redirect, useRouter } from "next/navigation";
+import { auth } from "~/server/auth"
 
 import { api } from "~/trpc/react";
 
@@ -46,11 +47,19 @@ interface Product {
   updatedAt: Date;
 }
 
-export default function ProfilePage({
+export default async function ProfilePage({
   params,
 }: {
   params: Promise<{ userId: string }>;
 }) {
+
+  const session = await auth()
+
+  // If user is not authenticated, redirect to login page
+  if (!session?.user) {
+    redirect("/login")
+  }
+
   const router = useRouter();
   const { userId } = use(params);
   const isUsername = userId.startsWith("%40");
