+++ src/app/profile/page.tsx
@@ -1,7 +1,17 @@
-import ProfilePage from "./profile-page-component";
+import { auth } from "~/server/auth"
+import { redirect } from "next/navigation"
+import ProfilePage from "./profile-page-component"
 
-export default function Page() {
-  return <ProfilePage />;
+export default async function Page() {
+  const session = await auth()
+
+  // If user is not authenticated, redirect to login page
+  if (!session?.user) {
+    redirect("/login")
+  }
+
+  return <ProfilePage />
 }
 
-export const dynamic = "force-dynamic";
+export const dynamic = "force-dynamic"
+
